api and test improvement for RuntimeObservableSetting
