+++ multiplatform-settings-runtime-observable/api/android/multiplatform-settings-runtime-observable.api
@@ -6,7 +6,6 @@ public final class com/russhwolf/settings/runtime_observable/BuildConfig {
 }
 
 public final class com/russhwolf/settings/serialization/ConverterKt {
-	public static final fun toRuntimeObservable (Lcom/russhwolf/settings/Settings;Z)Lcom/russhwolf/settings/ObservableSettings;
-	public static synthetic fun toRuntimeObservable$default (Lcom/russhwolf/settings/Settings;ZILjava/lang/Object;)Lcom/russhwolf/settings/ObservableSettings;
+	public static final fun toRuntimeObservable (Lcom/russhwolf/settings/Settings;)Lcom/russhwolf/settings/ObservableSettings;
 }
