+++ multiplatform-settings-runtime-observable/api/jvm/multiplatform-settings-runtime-observable.api
@@ -1,5 +1,4 @@
 public final class com/russhwolf/settings/serialization/ConverterKt {
-	public static final fun toRuntimeObservable (Lcom/russhwolf/settings/Settings;Z)Lcom/russhwolf/settings/ObservableSettings;
-	public static synthetic fun toRuntimeObservable$default (Lcom/russhwolf/settings/Settings;ZILjava/lang/Object;)Lcom/russhwolf/settings/ObservableSettings;
+	public static final fun toRuntimeObservable (Lcom/russhwolf/settings/Settings;)Lcom/russhwolf/settings/ObservableSettings;
 }
